adding calculation support TurboExpander.
