+++ src/propylean/equipments/generic_equipment_classes.py
@@ -1,11 +1,13 @@
 from propylean.equipments.abstract_equipment_classes import _EquipmentOneInletOutlet, _EquipmentMultipleInletOutlet
 import propylean.properties as prop
 from propylean.constants import Constants
+from propylean.settings import Settings
 from pandas import DataFrame
 from propylean.validators import _Validators
 from math import pi, sqrt, acos
+import fluids.compressible as compressible_fluid
 
-#Defining generic class for all types of pressure changers like Pumps, Compressors and Expanders.
+# Defining generic class for all types of pressure changers like Pumps, Compressors and Expanders.
 class _PressureChangers(_EquipmentOneInletOutlet):
     def __init__(self,**inputs) -> None:
         """ 
@@ -75,7 +77,6 @@ class _PressureChangers(_EquipmentOneInletOutlet):
         
         self.efficiency = 100 if 'efficiency' not in inputs else inputs['efficiency']
         
-        
     @property
     def suction_pressure(self):
         return super(_PressureChangers, self).inlet_pressure
@@ -97,6 +98,7 @@ class _PressureChangers(_EquipmentOneInletOutlet):
     @differential_pressure.setter
     def differential_pressure(self, value):
         _Validators.validate_arg_prop_value_type("differential_pressure", value, (prop.Pressure, int, float, tuple))
+        _Validators.validate_non_negative_value("differential_pressure", value)
         self = self._get_equipment_object(self)
         value, unit = self._tuple_property_value_unit_returner(value, prop.Pressure)
         if unit is None:
@@ -148,6 +150,214 @@ class _PressureChangers(_EquipmentOneInletOutlet):
             unit = self._power.unit         
         self._power = prop.Power(value, unit)
         self._update_equipment_object(self)  
+
+# Defining generic class for Compressors and Expanders.
+class _GasPressureChangers(_PressureChangers):
+    def __init__(self, **inputs) -> None:
+        """ 
+        DESCRIPTION:
+            class for creating objects to represent a GasPressureChanger.
+        
+        PARAMETERS:
+            Read _PressureChangers class for more arguments for this class  
+            efficiency:
+                Required: No
+                Type: int or float (recommended)
+                Acceptable values: Non-negative values
+                Default value: 1.4
+                Description: Efficiency of the compressor. Efficency can be set as
+                             adiabatic/isentropic or polytropic efficiency.
+            
+            adiabatic_efficiency:
+                Required: No
+                Type: int or float (recommended)
+                Acceptable values: Non-negative values
+                Default value: 0.7
+                Description: Efficiency of the compressor considering adiabatic/isentropic
+                             compression.
+            
+            polytropic_efficiency:
+                Required: No
+                Type: int or float (recommended)
+                Acceptable values: Non-negative values
+                Default value: 0.7
+                Description: Efficiency of the compressor considering polytropic
+                             compression.
+
+            polytropic_exponent:
+                Required: No
+                Type: int or float (recommended)
+                Acceptable values: Non-negative values
+                Default value: 1.4
+                Description: Polytropic exponent of the gas.
+        
+        PROPERTIES:
+            power:
+                Type: Power
+                Description: Power required by the compressor.
+
+        
+        RETURN VALUE:
+            Type: CentrifugalCompressor
+            Description: Returns an object of type CentrifugalCompressor with all properties of
+                         a Centrifugal Compressor used in process industries.
+        
+        ERROR RAISED:
+            Type:
+            Description:
+        
+        SAMPLE USE CASES:
+            >>> CC_1 = CentrifugalCompressor(tag="P1")
+            >>> print(CC_1)
+            Centrifugal Compressor with tag: P1
+        """
+        super().__init__( **inputs)
+        self.adiabatic_efficiency = 0.7 if 'efficiency' not in inputs else inputs['efficiency']
+        self.polytropic_exponent = 1.4 if 'polytropic_exponent' not in inputs else inputs['polytropic_exponent']
+        
+    @property
+    def temperature_change(self):
+        self = self._get_equipment_object(self)
+        k = self.polytropic_exponent
+        if (self._inlet_material_stream_index is not None or
+           self._outlet_material_stream_index is not None):
+            is_inlet = False if self._inlet_material_stream_index is None else True
+            k = self._connected_stream_property_getter(is_inlet, "material", "isentropic_exponent")
+            if Settings.compression_process == "Polytropic":
+                k = compressible_fluid.polytropic_exponent(k=k, eta_p=self.polytropic_efficiency)
+
+        T1 = self.inlet_temperature
+        P1 = self.inlet_pressure
+        P2 = self.outlet_pressure
+        T1.unit = "K"
+        P1.unit = P2.unit ="Pa"
+        eta = self.efficiency
+        value = prop.Temperature(compressible_fluid.isentropic_T_rise_compression(T1.value, P1.value, P2.value, k, eta))
+        value = value - T1
+        value.unit = self.inlet_temperature.unit
+        return value
+
+    @property
+    def efficiency(self):
+        self = self._get_equipment_object(self)
+        if Settings.compression_process == "Polytropic":
+            return self.polytropic_efficiency
+        else:
+            return self.adiabatic_efficiency
+    @efficiency.setter
+    def efficiency(self, value):
+        _Validators.validate_arg_prop_value_type("efficiency", value, (int, float))
+        self = self._get_equipment_object(self)
+        if value < 0:
+            raise Exception("Please enter a positive value for efficiency")
+        elif value <= 1:
+            value = value
+        else:
+            value = value/100
+        if Settings.compression_process.lower() in ["adiabatic", "isentropic"]:
+            self.adiabatic_efficiency = value
+        else:
+            self.polytropic_efficiency = value
+            
+        self._efficiency = value
+
+    @property
+    def adiabatic_efficiency(self):
+        self = self._get_equipment_object(self)
+        return self._adiabatic_efficiency
+    @adiabatic_efficiency.setter
+    def adiabatic_efficiency(self, value):
+        _Validators.validate_arg_prop_value_type("adiabatic_efficiency", value, (int, float))
+        self = self._get_equipment_object(self)
+        if value ==  None:
+            value = 0.7
+        self._adiabatic_efficiency = value
+        self._update_equipment_object(self)
+    
+    @property
+    def polytropic_efficiency(self):
+        self = self._get_equipment_object(self)
+        if (self._inlet_material_stream_index is not None or
+           self._outlet_material_stream_index is not None):
+            is_inlet = False if self._inlet_material_stream_index is None else True
+            isentropic_exponent = self._connected_stream_property_getter(is_inlet, "material", "isentropic_exponent")
+            return compressible_fluid.isentropic_efficiency(P1 = self._inlet_pressure.value,
+                                                            P2 = self._outlet_pressure.value,
+                                                            k = isentropic_exponent,
+                                                            eta_s = self.adiabatic_efficiency)
+        else:
+            return self.efficiency
+    @polytropic_efficiency.setter
+    def polytropic_efficiency(self, value):
+        _Validators.validate_arg_prop_value_type("polytropic_efficiency", value, (int, float))
+        self = self._get_equipment_object(self)
+        is_inlet = False if self._intlet_material_stream_index is None else True
+        isentropic_exponent = self._connected_stream_property_getter(is_inlet, "material", "isentropic_exponent")
+        self.adiabatic_efficiency = compressible_fluid.isentropic_efficiency(P1 = self._inlet_pressure.value,
+                                                                             P2 = self._outlet_pressure.value,
+                                                                             k = isentropic_exponent,
+                                                                             eta_p = value)
+        self._update_equipment_object(self)
+    
+    @property
+    def polytropic_exponent(self):
+        self = self._get_equipment_object(self)
+        if (self._inlet_material_stream_index is None and
+            self._outlet_material_stream_index is None):
+            return self._polytropic_exponent
+        is_inlet = False if self._inlet_material_stream_index is None else True
+        k = self._connected_stream_property_getter(is_inlet, "material", "isentropic_exponent")
+        return compressible_fluid.polytropic_exponent(k=k, eta_p=self.polytropic_efficiency)
+    @polytropic_exponent.setter
+    def polytropic_exponent(self, value):
+        _Validators.validate_arg_prop_value_type("polytropic_exponent", value, (int, float))
+        self = self._get_equipment_object(self)
+        if (self._inlet_material_stream_index is None and
+            self._outlet_material_stream_index is None):
+            self._polytropic_exponent = value
+        else:
+            raise Exception("""Polytropic Exponent cannot be set as Compressor is connected to a MaterialStream object.\n
+                               Update Isentropic Exponent of the stream object instead.""")
+
+    @property
+    def power(self):
+        self = self._get_equipment_object(self)
+        is_inlet = False if self._inlet_material_stream_index is None else True
+        isentropic_exponent = self._connected_stream_property_getter(is_inlet, "material", "isentropic_exponent")
+        Z = self._connected_stream_property_getter(is_inlet, "material", "Z_g")
+        MW = self._connected_stream_property_getter(is_inlet, "material", "molecular_weight")
+        work = compressible_fluid.isentropic_work_compression(T1 = self.inlet_temperature.value,
+                                                              k = isentropic_exponent,
+                                                              Z = Z,
+                                                              P1 = self._inlet_pressure.value,
+                                                              P2 = self._outlet_pressure.value,
+                                                              eta = self.adiabatic_efficiency)
+        return prop.Power(work * self.inlet_mass_flowrate.value / MW.value)
+    
+    @property
+    def head(self):
+        if (self._inlet_material_stream_index is None or
+            self._outlet_material_stream_index is None):
+            raise Exception("Head calculations only supported when Compressor is connected to a MaterialStream object.")
+        if Settings.compression_process.lower() in ["adiabatic", "isentropic"]:
+            isentropic_exponent = self._connected_stream_property_getter(True, "material", "isentropic_exponent")
+            ratio = (isentropic_exponent - 1)/isentropic_exponent
+        else:
+            ratio = (self.polytropic_exponent - 1)/self.polytropic_exponent
+        
+        P1 = self.inlet_pressure
+        P2 = self.outlet_pressure
+        P1.unit = "Pa"
+        P2.unit = "Pa"
+        Zi = self._connected_stream_property_getter(True, "material", "Z_g")
+        Zo = self._connected_stream_property_getter(False, "material", "Z_g")
+        Z = (Zi + Zo)/2
+        T1 = self.inlet_temperature
+        T1.unit = "K"
+        MW = self._connected_stream_property_getter(True, "material", "molecular_weight")
+        head = Z * Constants.R * T1.value * (pow((P2.value/P1.value), ratio) - 1)/(ratio * MW.value)
+        return prop.Length(head)
+        
     
 #Defining generic class for all types of vessels. 
 class _Vessels(_EquipmentOneInletOutlet):
